refactor: rename sample_feature → debug_feature
