+++ lib/src/app.dart
@@ -42,7 +42,7 @@ import 'package:reaprime/src/launcher/launcher_scan_page.dart';
 import 'package:reaprime/src/models/device/de1_interface.dart';
 import 'package:reaprime/src/models/device/scale.dart';
 import 'package:reaprime/src/plugins/plugin_loader_service.dart';
-import 'package:reaprime/src/sample_feature/scale_debug_view.dart';
+import 'package:reaprime/src/debug_feature/scale_debug_view.dart';
 import 'package:reaprime/src/services/storage/bean_storage_service.dart';
 import 'package:reaprime/src/services/storage/grinder_storage_service.dart';
 import 'package:reaprime/src/services/storage/profile_storage_service.dart';
@@ -54,9 +54,9 @@ import 'package:reaprime/src/settings/device_management_page.dart';
 import 'package:reaprime/src/settings/advanced_page.dart';
 import 'package:reaprime/src/settings/plugins_settings_view.dart';
 import 'package:reaprime/src/skin_selector/skin_selector_page.dart';
-import 'sample_feature/sample_item_details_view.dart';
+import 'debug_feature/sample_item_details_view.dart';
 
-import 'sample_feature/sample_item_list_view.dart';
+import 'debug_feature/sample_item_list_view.dart';
 import 'settings/gateway_mode.dart';
 import 'settings/settings_controller.dart';
 import 'settings/settings_view.dart';
