+++ lib/src/debug_feature/sample_item_details_view.dart
