+++ lib/src/debug_feature/sample_item_list_view.dart
@@ -4,8 +4,8 @@ import 'package:reaprime/src/controllers/device_controller.dart';
 import 'package:reaprime/src/models/device/de1_interface.dart';
 import 'package:reaprime/src/models/device/device.dart' as dev;
 import 'package:reaprime/src/models/device/scale.dart';
-import 'package:reaprime/src/sample_feature/sample_item_details_view.dart';
-import 'package:reaprime/src/sample_feature/scale_debug_view.dart';
+import 'package:reaprime/src/debug_feature/sample_item_details_view.dart';
+import 'package:reaprime/src/debug_feature/scale_debug_view.dart';
 import 'package:shadcn_ui/shadcn_ui.dart' hide Scale;
 
 /// Displays a list of discovered devices with Inspect and Connect actions.
